+++ packages/map-interface/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file. The format
 is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.8] - 2024-10-26
+
+- Fix inspector colors with new version of `chroma-js`
+- Introduce dependency on `@macrostrat/color-utils`
+
 ## [1.0.5] - 2024-10-24
 
 - Fix bundling